+++ tests/test_find_nearest_tests_for_pest.py
@@ -16,7 +16,7 @@
 # along with PHPUnitKit.  If not, see <https://www.gnu.org/licenses/>.
 
 from PHPUnitKit.tests import unittest
-from PHPUnitKit.plugin import find_nearest_tests
+from PHPUnitKit.lib.utils import find_nearest_tests
 
 
 class TestFindSelectedTestMethods(unittest.ViewTestCase):