+++ tests/test_phpunit.py
@@ -15,12 +15,12 @@
 # You should have received a copy of the GNU General Public License
 # along with PHPUnitKit.  If not, see <https://www.gnu.org/licenses/>.
 
-from PHPUnitKit.plugin import _get_phpunit_options
-from PHPUnitKit.plugin import set_last_run
+from PHPUnitKit.lib.utils import set_last_run
+from PHPUnitKit.lib.utils import get_phpunit_options
 from PHPUnitKit.tests import unittest
 
 
-@unittest.mock.patch.dict('PHPUnitKit.plugin._session', {'options': {}}, clear=True)
+@unittest.mock.patch.dict('PHPUnitKit.lib.utils._session', {'options': {}}, clear=True)
 class TestPHPUnitVisit(unittest.ViewTestCase):
 
     def setUp(self):
@@ -52,7 +52,7 @@ class TestPHPUnitVisit(unittest.ViewTestCase):
         open_file.assert_not_called()
 
 
-@unittest.mock.patch.dict('PHPUnitKit.plugin._session', {'options': {}}, clear=True)
+@unittest.mock.patch.dict('PHPUnitKit.lib.utils._session', {'options': {}}, clear=True)
 class TestPHPUnitToggle(unittest.ViewTestCase):
 
     def setUp(self):
@@ -61,49 +61,49 @@ class TestPHPUnitToggle(unittest.ViewTestCase):
 
     def test_toggle_boolean_with_no_options_or_session(self):
         self.run_window_command('phpunit_toggle_option', {'option': 'no-coverage'})
-        self.assertEquals({'no-coverage': True}, _get_phpunit_options(self.view))
+        self.assertEquals({'no-coverage': True}, get_phpunit_options(self.view))
         self.run_window_command('phpunit_toggle_option', {'option': 'no-coverage'})
-        self.assertEquals({'no-coverage': False}, _get_phpunit_options(self.view))
+        self.assertEquals({'no-coverage': False}, get_phpunit_options(self.view))
         self.run_window_command('phpunit_toggle_option', {'option': 'no-coverage', 'value': True})
-        self.assertEquals({'no-coverage': True}, _get_phpunit_options(self.view))
+        self.assertEquals({'no-coverage': True}, get_phpunit_options(self.view))
         self.run_window_command('phpunit_toggle_option', {'option': 'no-coverage', 'value': True})
-        self.assertEquals({'no-coverage': None}, _get_phpunit_options(self.view))
+        self.assertEquals({'no-coverage': None}, get_phpunit_options(self.view))
         self.run_window_command('phpunit_toggle_option', {'option': 'no-coverage', 'value': True})
-        self.assertEquals({'no-coverage': True}, _get_phpunit_options(self.view))
+        self.assertEquals({'no-coverage': True}, get_phpunit_options(self.view))
 
     def test_toggle_boolean_with_options_but_no_session(self):
         self.view.settings().set('phpunit.options', {'no-coverage': True})
         self.run_window_command('phpunit_toggle_option', {'option': 'no-coverage'})
-        self.assertEquals({'no-coverage': False}, _get_phpunit_options(self.view))
+        self.assertEquals({'no-coverage': False}, get_phpunit_options(self.view))
         self.run_window_command('phpunit_toggle_option', {'option': 'no-coverage'})
-        self.assertEquals({'no-coverage': True}, _get_phpunit_options(self.view))
+        self.assertEquals({'no-coverage': True}, get_phpunit_options(self.view))
         self.run_window_command('phpunit_toggle_option', {'option': 'no-coverage'})
-        self.assertEquals({'no-coverage': False}, _get_phpunit_options(self.view))
+        self.assertEquals({'no-coverage': False}, get_phpunit_options(self.view))
 
-    @unittest.mock.patch.dict('PHPUnitKit.plugin._session', {'options': {'no-coverage': True}}, clear=True)
+    @unittest.mock.patch.dict('PHPUnitKit.lib.utils._session', {'options': {'no-coverage': True}}, clear=True)
     def test_toggle_boolean_with_session(self):
         self.view.settings().set('phpunit.options', {})
         self.run_window_command('phpunit_toggle_option', {'option': 'no-coverage'})
-        self.assertEquals({'no-coverage': False}, _get_phpunit_options(self.view))
+        self.assertEquals({'no-coverage': False}, get_phpunit_options(self.view))
         self.run_window_command('phpunit_toggle_option', {'option': 'no-coverage'})
-        self.assertEquals({'no-coverage': True}, _get_phpunit_options(self.view))
+        self.assertEquals({'no-coverage': True}, get_phpunit_options(self.view))
         self.view.settings().set('phpunit.options', {'no-coverage': True})
         self.run_window_command('phpunit_toggle_option', {'option': 'no-coverage'})
-        self.assertEquals({'no-coverage': False}, _get_phpunit_options(self.view))
+        self.assertEquals({'no-coverage': False}, get_phpunit_options(self.view))
         self.view.settings().set('phpunit.options', {'no-coverage': False})
         self.run_window_command('phpunit_toggle_option', {'option': 'no-coverage'})
-        self.assertEquals({'no-coverage': True}, _get_phpunit_options(self.view))
+        self.assertEquals({'no-coverage': True}, get_phpunit_options(self.view))
 
     def test_toggle_string_with_no_options_or_session(self):
         self.run_window_command('phpunit_toggle_option', {'option': 'order-by=', 'value': 'depends,defects'})
-        self.assertEquals({'order-by=': 'depends,defects'}, _get_phpunit_options(self.view))
+        self.assertEquals({'order-by=': 'depends,defects'}, get_phpunit_options(self.view))
         self.run_window_command('phpunit_toggle_option', {'option': 'order-by=', 'value': 'depends,defects'})
-        self.assertEquals({'order-by=': None}, _get_phpunit_options(self.view))
+        self.assertEquals({'order-by=': None}, get_phpunit_options(self.view))
         self.run_window_command('phpunit_toggle_option', {'option': 'order-by=', 'value': 'depends,defects'})
-        self.assertEquals({'order-by=': 'depends,defects'}, _get_phpunit_options(self.view))
+        self.assertEquals({'order-by=': 'depends,defects'}, get_phpunit_options(self.view))
         self.run_window_command('phpunit_toggle_option', {'option': 'order-by=', 'value': 'default'})
-        self.assertEquals({'order-by=': 'default'}, _get_phpunit_options(self.view))
+        self.assertEquals({'order-by=': 'default'}, get_phpunit_options(self.view))
         self.run_window_command('phpunit_toggle_option', {'option': 'order-by=', 'value': 'defects'})
-        self.assertEquals({'order-by=': 'defects'}, _get_phpunit_options(self.view))
+        self.assertEquals({'order-by=': 'defects'}, get_phpunit_options(self.view))
         self.run_window_command('phpunit_toggle_option', {'option': 'order-by=', 'value': 'defects'})
-        self.assertEquals({'order-by=': None}, _get_phpunit_options(self.view))
+        self.assertEquals({'order-by=': None}, get_phpunit_options(self.view))